[refactor] 🐛 build 오류 수정 및  pages 네이밍 수정 - nextjs 에서 src > pages 폴더 에 있는 하위 파일들을 모두 page 로 인식하기 때문에 실제 page 만 인식하기 위해 파일 확장자 변경
