+++ src/pages/onboarding/index.page.tsx
